requirements.txt
google-generativeai>=0.8.5
requests>=2.32.5
python-dotenv>=1.1.1
flask>=2.3.0
aiohttp>=3.8.0
asyncio
